Move all admonitions to general directives directory
