+++ incubator/guides-restructured-text/src/RestructuredText/Directives/BestPracticeDirective.php
@@ -11,12 +11,12 @@ declare(strict_types=1);
  * @link https://phpdoc.org
  */
 
-namespace phpDocumentor\Guides\RestructuredText\HTML\Directives;
+namespace phpDocumentor\Guides\RestructuredText\Directives;
 
-class AdmonitionDirective extends AbstractAdmonitionDirective
+class BestPracticeDirective extends AbstractAdmonitionDirective
 {
     public function __construct()
     {
-        parent::__construct('admonition', 'Admonition');
+        parent::__construct('best-practice', 'Best Practice');
     }
 }
