+++ incubator/guides-restructured-text/src/RestructuredText/Directives/SeeAlsoDirective.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @link https://phpdoc.org
  */
 
-namespace phpDocumentor\Guides\RestructuredText\HTML\Directives;
+namespace phpDocumentor\Guides\RestructuredText\Directives;
 
 class SeeAlsoDirective extends AbstractAdmonitionDirective
 {
